+++ src/js/extend.js
@@ -1,6 +1,10 @@
 'use strict';
 // Extend native objects, aka monkey patching ..... really I see no harm!
 
+// This require is just there so Alex and other XPath-evaluator-replacers get an automatic notification to extend the date object.
+// It is not required for those that use enketo-xpathjs
+require( 'enketo-xpathjs/src/date-extensions' );
+
 /**
  * Pads a string with prefixed zeros until the requested string length is achieved.
  * @param  {number} digits [description]
@@ -14,43 +18,7 @@ String.prototype.pad = function( digits ) {
     return x;
 };
 
-/**
- * Converts a native Date UTC String to a RFC 3339-compliant date string with local offsets
- * used in JavaRosa, so it replaces the Z in the ISOstring with a local offset
- * @return {string} a datetime string formatted according to RC3339 with local offset
- */
-Date.prototype.toISOLocalString = function() {
-    //2012-09-05T12:57:00.000-04:00 (ODK)
 
-    if ( this.toString() === 'Invalid Date' ) {
-        return this.toString();
-    }
-
-    return new Date( this.getTime() - ( this.getTimezoneOffset() * 60 * 1000 ) ).toISOString()
-        .replace( 'Z', this.getTimezoneOffsetAsTime() );
-};
-
-Date.prototype.getTimezoneOffsetAsTime = function() {
-    var offsetMinutesTotal;
-    var hours;
-    var minutes;
-    var direction;
-    var pad2 = function( x ) {
-        return ( x < 10 ) ? '0' + x : x;
-    };
-
-    if ( this.toString() === 'Invalid Date' ) {
-        return this.toString();
-    }
-
-    offsetMinutesTotal = this.getTimezoneOffset();
-
-    direction = ( offsetMinutesTotal < 0 ) ? '+' : '-';
-    hours = pad2( Math.abs( Math.floor( offsetMinutesTotal / 60 ) ) );
-    minutes = pad2( Math.abs( Math.floor( offsetMinutesTotal % 60 ) ) );
-
-    return direction + hours + ':' + minutes;
-};
 
 if ( typeof console.deprecate === 'undefined' ) {
     console.deprecate = function( bad, good ) {